Refactored quiz and question controllers and models
